requirements.txt
aiogram==3.27.0
deezer-python==7.3.0
aiohttp==3.9.5
